% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCDQR_em.R
\name{ABCDQR_em}
\alias{ABCDQR_em}
\title{em algorithm}
\usage{
ABCDQR_em(y, u, Ai, Bi, Ci, Di, Qi, Ri, m1i, P1i, nx, ny, nu, Ae = TRUE,
  Be = TRUE, Ce = TRUE, De = TRUE, Qe = TRUE, Re = TRUE,
  m1e = TRUE, P1e = TRUE, max_iter = 100, tol = 1e-06,
  txo = FALSE)
}
\arguments{
\item{y:}{data. Matrix ny*nt
A: matrix nx*nx
C: matrix ny*nx}
}
\value{
y by rows, ny = nrow(y), nt = ncol(y)
}
\description{
Kalman filter for state space model
}
