% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCDQR_kfilter.R
\name{ABCDQR_kfilter}
\alias{ABCDQR_kfilter}
\title{Kalman filter}
\usage{
ABCDQR_kfilter(y, u, A, B, C, D, Q, R, x10, P10, nx, ny, nu)
}
\arguments{
\item{y:}{data. Matrix ny*nt
A: matrix nx*nx
C: matrix ny*nx}
}
\value{
y by rows, ny = nrow(y), nt = ncol(y)
}
\description{
Kalman filter for state space model
}
