% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCDQR_predict.R
\name{ABCDQR_predict}
\alias{ABCDQR_predict}
\title{forecasting function}
\usage{
ABCDQR_predict(
  y,
  u,
  A,
  B,
  C,
  D,
  Q,
  R,
  m1,
  P1,
  nx,
  ny,
  nu,
  u_new,
  n_ahead,
  conf_level = 0.95
)
}
\arguments{
\item{y}{data}
}
\value{
y by rows, m = nrow(y), nt = ncol(y)
}
\description{
forecasting function
}
