% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACQR_em.R
\name{ACQR_em}
\alias{ACQR_em}
\title{em algorithm}
\usage{
ACQR_em(y, Ai, Ci, Qi, Ri, m1i, P1i, nx, ny, Ae = TRUE, Ce = TRUE,
  Qe = TRUE, Re = TRUE, m1e = TRUE, P1e = TRUE, max_iter = 100,
  tol = 1e-06, txo = FALSE)
}
\arguments{
\item{y:}{data. Matrix ny*nt
A: matrix nx*nx
C: matrix ny*nx}
}
\value{
y by rows, ny = nrow(y), nt = ncol(y)
}
\description{
Kalman filter for state space model
}
