% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACQR_em2.R
\name{ACQR_em2}
\alias{ACQR_em2}
\title{em algorithm}
\usage{
ACQR_em2(
  y,
  Ai,
  Ci,
  Qi,
  Ri,
  m1i,
  P1i,
  nx,
  ny,
  Ae = matrix(1, nx, nx),
  Ce = matrix(1, ny, nx),
  Qe = matrix(1, nx, nx),
  Re = matrix(1, ny, ny),
  m1e = matrix(1, nx, 1),
  P1e = matrix(1, nx, nx),
  max_iter = 100,
  tol = 1e-06,
  txo = FALSE
)
}
\arguments{
\item{y:}{data. Matrix ny*nt
A: matrix nx*nx
C: matrix ny*nx}
}
\value{
y by rows, ny = nrow(y), nt = ncol(y)
}
\description{
Kalman filter for state space model
}
