% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACQR_ksmoother.R
\name{ACQR_ksmoother}
\alias{ACQR_ksmoother}
\title{Kalman smoother}
\usage{
ACQR_ksmoother(A, xtt, Ptt, xtt1, Ptt1)
}
\arguments{
\item{y}{output data}
}
\value{
y by rows, ny = nrow(y), nt = ncol(y)
}
\description{
Kalman smoother for state space model
  x_{t+1} = Ax_{t} + w_{t}
  y_{t}   = Cx_{t} + v_{t}
}
