% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACQRS_sub.R
\name{ACQRS_sub}
\alias{ACQRS_sub}
\title{Estimate the state space model using subspace algorithm}
\usage{
ACQRS_sub(y, nx, ny, i = nx + 1)
}
\arguments{
\item{\bold{y}}{data}

\item{\bold{nx}}{number of rows of matrix A}

\item{\bold{i}}{auxiliar parameter for building the Hankel matrix. By default, i = nx+1}
}
\value{
A, C, Q, R, S
}
\description{
Subspace algorithm for the estimation of model
\deqn{x_{t+1} = Ax_{t} + w_{t}}
\deqn{y_{t} = Cx_{t} + v_{t}}
where \eqn{w_{t}} and \eqn{v_{t}} have zero mean and covariance matrices
\deqn{Var(w_{t})=Q, Var(v_{t})=R, Cov(w_{t},v_{t})=S}
}
\references{
\emph{Subspace Identification for Linear Systems.}
 \emph{Theory - Implementation - Applications.}
 Peter Van Overschee / Bart De Moor.
 Kluwer Academic Publishers, 1996
}
