% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hankel_yij.R
\name{hankel_yij}
\alias{hankel_yij}
\title{Build a Hankel matrix}
\usage{
hankel_yij(y, i, j)
}
\arguments{
\item{y:}{data matrix (m x N), i: number of row blocks, j: number of columns}
}
\value{
H in R^\{m*i x j\}
}
\description{
Hankel matrix with i rows and j columns from y
}
\examples{

}
\keyword{Hankel}
\keyword{matrix}
