% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACQR_predict.R
\name{ACQR_predict}
\alias{ACQR_predict}
\title{forecasting function}
\usage{
ACQR_predict(y, A, C, Q, R, m1, P1, nx, ny, n_ahead = 0, conf_level = 0.95)
}
\arguments{
\item{y}{data}
}
\value{
y by rows, m = nrow(y), nt = ncol(y)
}
\description{
forecasting function for state space model
  x_{t+1} = Ax_{t} + w_{t}
  y_{t}   = Cx_{t} + v_{t}
  cov(w_{t},v_{t}) = [Q 0;0 R]
}
